Move two-factor management into settings page
